refactor: rename test files and update namespaces to Apizr
